requirements.txt
asgiref==3.2.10
boto3==1.16.7
botocore==1.19.7
Django==3.1.2
django-cors-headers==3.5.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
jmespath==0.10.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
typing==3.7.4.3
urllib3==1.25.11
